fixing conflits

+++ src/main/java/com/gladguys/alucoapi/repositories/customs/impl/CustomClassRepositoryImpl.java
@@ -12,6 +12,7 @@ import org.springframework.stereotype.Repository;
 import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -35,12 +36,13 @@ public class CustomClassRepositoryImpl implements CustomClassRepository {
 
 	@Override
 	public void deleteStudentFromClass(Long studentId, Long classId, Set<Long> examsId) {
-		NamedParameterJdbcTemplate template = new NamedParameterJdbcTemplate(this.jdbcTemplate.getDataSource());
+		NamedParameterJdbcTemplate template = new NamedParameterJdbcTemplate(Objects.requireNonNull(this.jdbcTemplate.getDataSource()));
 		MapSqlParameterSource parameters = new MapSqlParameterSource();
 		parameters.addValue("studentId", studentId);
-		parameters.addValue("examsId", examsId);
-
-		template.update("DELETE FROM grade WHERE student_id = :studentId AND exam_id IN (:examsId) ", parameters);
+		if(examsId.size()!= 0) {
+			parameters.addValue("examsId", examsId);
+			template.update("DELETE FROM grade WHERE student_id = :studentId AND exam_id IN (:examsId) ", parameters);
+		}
 		this.jdbcTemplate.update("DELETE FROM student_class WHERE student_id = ? AND class_id = ?", new Object[]{studentId, classId});
 	}
 }
